feat: 원화 표시 제한설정 & 컴포넌트 분리
