+++ src/App.tsx
@@ -1,53 +1,43 @@
-import React, { ChangeEvent, useEffect, useRef, useState } from "react";
+import React, { ChangeEvent, useEffect, useState } from "react";
 import { getEurInfo } from './api/exchange';
 import { ExchangeInfo } from './model/Model';
+import { ExchangeInput } from './view/component/Info/ExchangeInput';
+import { ExchangeTotalInfo } from './view/component/Info/ExchangeTotalInfo';
 
 
 export const App = () => {
   const [eurInfo, setEurInfo] = useState<ExchangeInfo | undefined>(undefined);
-  const [eurInput, setEurInput] = useState<string>();
-  const [krw, setkrw] = useState<string>();
+  // const [eurInput, setEurInput] = useState<number>();
+  // const [krw, setkrw] = useState<string>('0');
 
-  const getEurInput = (e :ChangeEvent<HTMLInputElement>) => {
-    setEurInput(e.target.value);
-    // let eurInput :string = e.target.value;
-    console.log('eurInput:',eurInput)
-  }
-  // const exchangeEurToKrw = (krw: any) => krw * eurInfo.basePrice;
+
+  // const getEurInput = (e :ChangeEvent<HTMLInputElement>) => {
+  //     setEurInput(parseFloat(e.target.value));
+  // }
 
   useEffect(() => {
     getEurInfo().then((priceValue) => { setEurInfo(priceValue) });
     return () => {}
   }, []);
 
-  useEffect(() => {
-    if (eurInput != undefined && eurInfo?.basePrice) {
-      let totalPrice = parseFloat(eurInput) * eurInfo.basePrice;
-      setkrw(totalPrice.toLocaleString('ko-KR'));
-    }
-  },[eurInput]);
+  // if(krw == 'NaN') setkrw('0')
 
+  // useEffect(() => {
+  //   if (eurInput != undefined && eurInfo?.basePrice) {
+  //     let totalPrice :number = Math.ceil(eurInput * eurInfo.basePrice);
+  //     setkrw(totalPrice.toLocaleString('ko-KR'));
+  //   }
 
-  console.log('eurInfo:',eurInfo);
+  // },[eurInput]);
+
+  
   if (!eurInfo) return null;
   return (
     <div className="App">
-      <div>환율기준 (1 유로)</div>
-      <div>
-        {eurInfo.basePrice}
-        {eurInfo.basePrice - eurInfo.openingPrice > 0 && "▲"}
-        {eurInfo.basePrice - eurInfo.openingPrice < 0 && "▼"}
-        {eurInfo.changePrice}원 (
-        {(eurInfo.changePrice / eurInfo.basePrice) * 100}%)
-      </div>
-      <div>
-        <div>살때 : {eurInfo.cashBuyingPrice}</div>
-        <div>팔때 : {eurInfo.cashSellingPrice}</div>
-        <div>보낼때 : {eurInfo.ttSellingPrice}</div>
-        <div>받을때 : {eurInfo.ttBuyingPrice}</div>
-      </div>
+      <ExchangeTotalInfo eurInfo={eurInfo}/>
       <hr />
-      <input onChange={getEurInput}/> 유로 ▶︎ <input disabled value={krw}/> 원
+      <ExchangeInput eurInfo={eurInfo}/>
+      {/* <input onChange={getEurInput}/> 유로 ▶︎ <input disabled value={krw}/> 원 */}
     </div>
   );
 };
