+++ tsconfig.json
@@ -23,5 +23,5 @@
   },
   "include": [
     "src"
-  ]
+, "index.tsx", "App.tsx"  ]
 }
